+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to Fuse will be documented in this file.
 
+## 2025-05-12
+
+### Added
+
+- Add user public SSH Keys
+    - user ssh keys are added to all new server provisions
+    - adding a new ssh key adds the key to existing provisioned servers
+
 ## 2025-05-11
 
 ### Added